refactor(ai-engine): resolve #1768 — split diffusion_lora.py into focused modules
